pyproject.toml
[tool.poetry]
name = "count-crab"
version = "0.1.0"
description = ""
authors = ["Isaman sangbumrung  <38631981+tao-Isaman@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "count_crab"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
line-bot-sdk = "^2.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
